+++ speedd-ml/src/main/scala/org/speedd/ml/learners/Learner.scala
@@ -3,5 +3,5 @@ package org.speedd.ml.learners
 import auxlib.log.Logging
 
 trait Learner extends Logging {
-  def trainFor(startTs: Long, endTs: Long, batchSize: Int): Unit
+  def trainFor(startTs: Int, endTs: Int, batchSize: Int): Unit
 }
